synfutures v1 tvl: load pairs from graph
